+++ src/components/Loginon/LogonSecond.jsx
@@ -1,28 +1,47 @@
 import { useState, useEffect } from "react";
 import styled from "styled-components";
-
-import RBarrow from "../../images/icons/RBarrow_logo.svg";
-import logon_check from "../../images/icons/logon_check.svg";
-import LogonModal from "./LogonModal";
-
 import logonStore from "../../stores/LogonStore";
-import {
-  createUserWithEmailAndPassword,
-  updateProfile,
-  signOut,
-} from "firebase/auth";
-// import { getScrollbarWidth } from "../../util";
-import { doc, setDoc } from "firebase/firestore";
-import { auth, db } from "../../firebase";
-import DaumPostcode from "react-daum-postcode";
-import useBodyScrollLock from "../../hook/useBodyScrollLock";
 
-// styled 부분
-const Form = styled.form`
+const Input = styled.input`
+  width: 100%;
+  height: 70px;
+  border: 1px solid var(--grayC);
+  border-radius: 4px;
+  font-size: 1.6rem;
+  font-family: "Figtree", "Pretendard", sans-serif;
+  padding: 15px;
+  &:focus {
+    outline: none;
+    border: 1px solid var(--dark);
+  }
+  &::placeholder {
+    font-size: 1.6rem;
+    color: var(--grayC);
+  }
+  @media screen and (max-width: 1024px) {
+    height: 56px;
+    font-size: 1.6rem;
+    &::placeholder {
+      font-size: 1.6rem;
+    }
+  }
+  @media screen and (max-width: 600px) {
+    height: 44px;
+    font-size: 1.4rem;
+    &::placeholder {
+      font-size: 1.4rem;
+    }
+  }
+`;
+
+const AllInputWrapper = styled.div`
   width: 100%;
   display: flex;
   flex-direction: column;
-  gap: 40px;
+  gap: 10px;
+  @media screen and (max-width: 600px) {
+    gap: 8px;
+  }
 `;
 
 const SubTWrapper = styled.div`
@@ -30,32 +49,123 @@ const SubTWrapper = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: start;
-  gap: 15px;
-  @media screen and (max-width: 1024px) {
-    gap: 12px;
-  }
+  gap: 5px;
   @media screen and (max-width: 600px) {
-    gap: 10px;
+    gap: 4px;
   }
 `;
 
 const Subsubtitle = styled.h4`
-  font-size: 2rem;
-  font-weight: bold;
-  span {
-    font-weight: 400;
-    color: var(--gray8);
-  }
+  font-size: 1.4rem;
+  font-weight: 500;
   @media screen and (max-width: 1024px) {
-    font-size: 1.6rem;
+    font-size: 1.2rem;
   }
   @media screen and (max-width: 600px) {
-    font-size: 1.2rem;
+    font-size: 0.9rem;
   }
 `;
 
-const Input = styled.input`
+const StyledSelect = styled.select`
   width: 100%;
+  height: 70px;
+  border: 1px solid var(--grayC);
+  border-radius: 4px;
+  font-size: 1.6rem;
+  font-family: "Figtree", "Pretendard", sans-serif;
+  padding-left: 15px;
+  background: ${({ $isTeamPlaceholder }) =>
+    $isTeamPlaceholder ? "var(--grayF5)" : "var(--light)"};
+
+  color: ${({ $isTeamPlaceholder }) =>
+    $isTeamPlaceholder ? "var(--grayC)" : "var(--dark)"};
+  appearance: none;
+  background: white
+    url('data:image/svg+xml;utf8,<svg fill="gray" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M7 10l5 5 5-5z"/></svg>')
+    no-repeat right 10px center;
+  background-size: 16px;
+  &:focus {
+    outline: none;
+    border: 1px solid var(--dark);
+  }
+  &::placeholder,
+  option {
+    font-size: 1.6rem;
+    color: black;
+    background: var(--grayF5);
+  }
+  @media screen and (max-width: 1024px) {
+    height: 56px;
+    font-size: 1.6rem;
+    &::placeholder,
+    option {
+      font-size: 1.6rem;
+    }
+  }
+  @media screen and (max-width: 600px) {
+    height: 44px;
+    font-size: 1.2rem;
+    &::placeholder,
+    option {
+      font-size: 1.2rem;
+    }
+  }
+`;
+
+const StyledSelect2Wrapper = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+`;
+
+const StyledSelect2 = styled.select`
+  width: 32%;
+  height: 70px;
+  border: 1px solid var(--grayC);
+  border-radius: 4px;
+  font-size: 1.6rem;
+  font-family: "Figtree", "Pretendard", sans-serif;
+  padding-left: 15px;
+  background: ${({ $isBirthPlaceholder }) =>
+    $isBirthPlaceholder ? "var(--grayF5)" : "var(--light)"};
+
+  color: ${({ $isBirthPlaceholder }) =>
+    $isBirthPlaceholder ? "var(--grayC)" : "var(--dark)"};
+  appearance: none;
+  background: white
+    url('data:image/svg+xml;utf8,<svg fill="gray" height="24" viewBox="0 0 24 24" width="24" xmlns="http://www.w3.org/2000/svg"><path d="M7 10l5 5 5-5z"/></svg>')
+    no-repeat right 10px center;
+  background-size: 16px;
+  &:focus {
+    outline: none;
+    border: 1px solid var(--dark);
+  }
+  &::placeholder,
+  option {
+    font-size: 1.6rem;
+    color: black;
+    background: var(--grayF5);
+  }
+  @media screen and (max-width: 1024px) {
+    height: 56px;
+    font-size: 1.6rem;
+    &::placeholder,
+    option {
+      font-size: 1.6rem;
+    }
+  }
+  @media screen and (max-width: 600px) {
+    height: 44px;
+    font-size: 1.2rem;
+    &::placeholder,
+    option {
+      font-size: 1.2rem;
+    }
+  }
+`;
+
+const Input2 = styled.input`
+  width: 32%;
   height: 70px;
   border: 1px solid var(--grayC);
   border-radius: 4px;
@@ -86,147 +196,16 @@ const Input = styled.input`
   }
 `;
 
-const PostWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-`;
-
-const PostInput = styled.input`
-  width: 60%;
-  height: 70px;
-  border: 1px solid var(--grayC);
-  border-radius: 4px;
-  font-size: 1.6rem;
-  padding: 15px;
-  &:focus {
-    outline: none;
-    border: 1px solid var(--dark);
-  }
-  &::placeholder,
-  input {
-    font-size: 1.6rem;
-    color: var(--grayC);
-  }
-  @media screen and (max-width: 1024px) {
-    height: 50px;
-    font-size: 1.6rem;
-    &::placeholder,
-    input {
-      font-size: 1.6rem;
-    }
-  }
-  @media screen and (max-width: 600px) {
-    height: 40px;
-    font-size: 1.2rem;
-    &::placeholder,
-    input {
-      font-size: 1.2rem;
-    }
-  }
-`;
-
-const PostButton = styled.button`
-  width: 38%;
-  height: 70px;
-  border: 1px solid var(--dark);
-  background: var(--dark);
-  color: var(--light);
-  border-radius: 4px;
-  font-size: 1.6rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-  transition: background-color 0.3s;
-  &:hover {
-    background: var(--gray2);
-  }
-  @media screen and (max-width: 1024px) {
-    height: 50px;
-    font-size: 1.4rem;
-  }
-  @media screen and (max-width: 600px) {
-    height: 40px;
-    font-size: 1.2rem;
-  }
-`;
-
-const Line = styled.span`
-  display: inline-block;
-  width: 100%;
-  height: 1px;
-  background: var(--dark);
-`;
-
-const AllCheckWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  @media screen and (max-width: 1024px) {
-    gap: 16px;
-  }
-  @media screen and (max-width: 600px) {
-    gap: 12px;
-  }
-`;
-
-const CheckWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  cursor: pointer;
-  padding-right: 10px;
-  @media screen and (max-width: 1024px) {
-    padding-right: 8px;
-  }
-  @media screen and (max-width: 600px) {
-    padding-right: 6px;
-  }
-`;
-
-const Checkoption = styled.div`
-  display: flex;
-  justify-content: start;
-  gap: 5px;
-`;
-
-const CheckCircle = styled.div`
-  width: 14px;
-  height: 14px;
-  border: 1px solid
-    ${({ checked }) => (checked ? "var(--dark)" : "var(--grayC)")};
-  border-radius: 50%;
-  position: relative;
-  img {
-    position: absolute;
-    left: 1px;
-    display: ${({ checked }) => (checked ? "block" : "none")};
-  }
-`;
-
-const CheckText = styled.h5`
-  font-size: 1.6rem;
-  span {
-    color: var(--red);
-  }
-  @media screen and (max-width: 1024px) {
-    font-size: 1.2rem;
-  }
-  @media screen and (max-width: 500px) {
-    font-size: 1rem;
-  }
-`;
-
 const LoginBtn = styled.button`
   width: 100%;
   height: 70px;
+  background: var(--grayE);
   border: none;
   border-radius: 4px;
   font-size: 2.4rem;
   color: var(--light);
   cursor: pointer;
+  margin-top: 30px;
   background: ${({ $valid }) => ($valid ? "var(--dark)" : "var(--grayE)")};
   transition: background-color 0.3s;
   &:hover {
@@ -254,405 +233,209 @@ const ErrorMessage = styled.p`
   }
 `;
 
-const ModalOverlay = styled.div`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-`;
+const currentyear = new Date().getFullYear();
+const years = Array.from(
+  { length: currentyear - 1920 + 1 },
+  (_, i) => currentyear - i
+);
+const months = Array.from({ length: 12 }, (_, i) => i + 1);
+const dates = Array.from({ length: 31 }, (_, i) => i + 1);
 
-const ModalContent = styled.div`
-  background: var(--light);
-  width: 500px;
-  padding: 20px;
-  border-radius: 8px;
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-  @media screen and (max-width: 600px) {
-    width: 90%;
-  }
-`;
-
-const ModalCloseButton = styled.button`
-  align-self: flex-end;
-  background: none;
-  border: none;
-  font-size: 1.6rem;
-  cursor: pointer;
-`;
-// styled 부분
+//
 
 const LogonSecond = () => {
   const { formData, setFormData, nextStep } = logonStore();
   const [errors, setErrors] = useState({});
-  const [isLoading, setIsLoading] = useState(false);
   const [isFormValid, setIsFormValid] = useState(false);
-  const [modalState, setModalState] = useState({
-    required: false,
-    privacy: false,
-  });
-  const [isAddressModalOpen, setIsAddressModalOpen] = useState(false);
 
-  const handleCheck = (key) => {
-    setFormData({
-      agreements: {
-        ...formData.agreements,
-        [key]: !formData.agreements[key],
-        ...(key === "all" && {
-          required: !formData.agreements.all,
-          privacy: !formData.agreements.all,
-          promotion: !formData.agreements.all,
-        }),
-        ...(key !== "all" && {
-          all:
-            (key === "required"
-              ? !formData.agreements.required
-              : formData.agreements.required) &&
-            (key === "privacy"
-              ? !formData.agreements.privacy
-              : formData.agreements.privacy) &&
-            (key === "promotion"
-              ? !formData.agreements.promotion
-              : formData.agreements.promotion),
-        }),
-      },
-    });
-  };
-
-  const handleSubmit = async (e) => {
+  const handleSubmit = (e) => {
     e.preventDefault();
     const newErrors = {};
 
-    // 유효성 검사
-    if (!formData.email) {
-      newErrors.email = "이메일을 입력해주세요.";
-    } else if (!/\S+@\S+\.\S+/.test(formData.email)) {
-      newErrors.email = "유효한 이메일 형식이 아닙니다.";
+    // 검증
+    if (!formData.username) {
+      newErrors.username = "이름을 입력해주세요.";
     }
-
-    if (!formData.password) {
-      newErrors.password = "비밀번호를 입력해주세요.";
-    } else if (formData.password.length < 6 || formData.password.length > 16) {
-      newErrors.password = "비밀번호는 6자 이상 16자 이하로 입력해주세요.";
+    if (!formData.favoriteTeam) {
+      newErrors.favoriteTeam = "응원 팀을 선택해주세요.";
     }
-
-    if (formData.password !== formData.confirmPassword) {
-      newErrors.confirmPassword = "비밀번호가 일치하지 않습니다.";
+    if (
+      !formData.birthdate.year ||
+      !formData.birthdate.month ||
+      !formData.birthdate.date
+    ) {
+      newErrors.birthdate = "생년월일을 모두 선택해 주세요.";
     }
-
-    if (!formData.nickname) {
-      newErrors.nickname = "닉네임을 입력해주세요.";
-    }
-
-    if (!formData.agreements.required || !formData.agreements.privacy) {
-      newErrors.agreements = "필수 약관에 동의해야 합니다.";
+    if (
+      !formData.phoneNumber.part1 ||
+      !formData.phoneNumber.part2 ||
+      !formData.phoneNumber.part3
+    ) {
+      newErrors.phoneNumber = "휴대폰 번호를 모두 입력해주세요.";
+    } else if (
+      !/^\d{3}$/.test(formData.phoneNumber.part1) ||
+      !/^\d{4}$/.test(formData.phoneNumber.part2) ||
+      !/^\d{4}$/.test(formData.phoneNumber.part3)
+    ) {
+      newErrors.phoneNumber = "유효한 휴대폰 번호를 입력해주세요.";
     }
 
     if (Object.keys(newErrors).length > 0) {
       setErrors(newErrors);
       return;
     }
-
-    setIsLoading(true);
     setErrors({});
-
-    try {
-      // Firebase Auth 사용자 생성
-      const userCredential = await createUserWithEmailAndPassword(
-        auth,
-        formData.email,
-        formData.password
-      );
-
-      const user = userCredential.user;
-
-      // 사용자 displayName 업데이트
-      await updateProfile(user, {
-        displayName: formData.username,
-      });
-
-      // Firestore에 사용자 정보 저장
-      await setDoc(doc(db, "users", user.uid), {
-        uid: user.uid,
-        username: formData.username,
-        favoriteTeam: formData.favoriteTeam,
-        birthdate: `${
-          formData.birthdate.year
-        }-${formData.birthdate.month.padStart(
-          2,
-          "0"
-        )}-${formData.birthdate.date.padStart(2, "0")}`,
-        phoneNumber: `${formData.phoneNumber.part1}-${formData.phoneNumber.part2}-${formData.phoneNumber.part3}`,
-        nickname: formData.nickname,
-        email: formData.email,
-        postalCode: formData.postalCode, // 우편번호 저장
-        address: formData.address, // 주소 저장
-        detailedAddress: formData.detailedAddress, // 상세주소 저장
-        createdAt: new Date().toISOString().split("T")[0],
-      });
-
-      await signOut(auth);
-      nextStep(); // → 가입 완료 페이지로 이동
-    } catch (error) {
-      if (error.code === "auth/email-already-in-use") {
-        setErrors({ email: "이미 가입된 이메일입니다." });
-      } else {
-        setErrors({ firebase: "회원가입 중 오류가 발생했습니다." });
-      }
-    } finally {
-      setIsLoading(false);
-    }
+    nextStep();
   };
 
   const handleChange = (e) => {
     const { name, value } = e.target;
-    setFormData({ [name]: value });
-  };
-
-  const handleAddressComplete = (data) => {
-    let fullAddress = data.address; // 도로명 주소
-    let extraAddress = "";
-
-    if (data.addressType === "R") {
-      // 도로명 주소인 경우
-      if (data.bname !== "") {
-        extraAddress += data.bname;
-      }
-      if (data.buildingName !== "") {
-        extraAddress +=
-          extraAddress !== "" ? `, ${data.buildingName}` : data.buildingName;
-      }
-      fullAddress += extraAddress !== "" ? ` (${extraAddress})` : "";
+    if (name === "username" || name === "favoriteTeam") {
+      setFormData({ [name]: value });
+    } else if (name.startsWith("birthdate")) {
+      const field = name.split(".")[1];
+      setFormData({
+        birthdate: { ...formData.birthdate, [field]: value },
+      });
+    } else if (name.startsWith("phoneNumber")) {
+      const field = name.split(".")[1];
+      setFormData({
+        phoneNumber: { ...formData.phoneNumber, [field]: value },
+      });
     }
-
-    setFormData({
-      postalCode: data.zonecode, // 우편번호
-      address: fullAddress, // 도로명 주소 + 추가 정보
-      detailedAddress: "", // 상세주소는 사용자가 입력
-    });
-
-    setIsAddressModalOpen(false); // 모달 닫기
-  };
-
-  const openModal = (type) => {
-    setModalState((prev) => ({ ...prev, [type]: true }));
-  };
-
-  const closeModal = (type) => {
-    setModalState((prev) => ({ ...prev, [type]: false }));
-  };
-
-  const openAddressModal = () => {
-    setIsAddressModalOpen(true);
-  };
-
-  const closeAddressModal = () => {
-    setIsAddressModalOpen(false);
   };
 
   useEffect(() => {
-    const isValid =
-      formData.email &&
-      /\S+@\S+\.\S+/.test(formData.email) &&
-      formData.password &&
-      formData.password.length >= 6 &&
-      formData.password.length <= 16 &&
-      formData.password === formData.confirmPassword &&
-      formData.nickname &&
-      formData.agreements.required &&
-      formData.agreements.privacy;
+    const isFormValid =
+      formData.username &&
+      formData.favoriteTeam &&
+      formData.birthdate.year &&
+      formData.birthdate.month &&
+      formData.birthdate.date &&
+      /^\d{3}$/.test(formData.phoneNumber.part1) &&
+      /^\d{4}$/.test(formData.phoneNumber.part2) &&
+      /^\d{4}$/.test(formData.phoneNumber.part3);
 
-    setIsFormValid(isValid);
+    setIsFormValid(isFormValid);
   }, [formData]);
 
-  //스크롤 막기
-  useBodyScrollLock(isAddressModalOpen);
-
   return (
-    <Form onSubmit={handleSubmit}>
-      <SubTWrapper>
-        <Subsubtitle>필수 입력란</Subsubtitle>
-        <Input
-          type="email"
-          name="email"
-          value={formData.email}
-          onChange={handleChange}
-          placeholder="이메일"
-        />
-        {errors.email && <ErrorMessage>{errors.email}</ErrorMessage>}
-        <Input
-          type="password"
-          name="password"
-          value={formData.password}
-          onChange={handleChange}
-          placeholder="비밀번호"
-        />
-        {errors.password && <ErrorMessage>{errors.password}</ErrorMessage>}
-        <Input
-          type="password"
-          name="confirmPassword"
-          value={formData.confirmPassword}
-          onChange={handleChange}
-          placeholder="비밀번호 재입력"
-        />
-        {errors.confirmPassword && (
-          <ErrorMessage>{errors.confirmPassword}</ErrorMessage>
-        )}
+    <form onSubmit={handleSubmit}>
+      <AllInputWrapper>
         <Input
           type="text"
-          name="nickname"
-          value={formData.nickname}
+          name="username"
+          value={formData.username}
           onChange={handleChange}
-          placeholder="닉네임"
+          placeholder="이름"
         />
-        {errors.nickname && <ErrorMessage>{errors.nickname}</ErrorMessage>}
-      </SubTWrapper>
-      <Line />
-      <SubTWrapper>
-        <Subsubtitle>
-          우편번호 <span>(선택)</span>
-        </Subsubtitle>
-        <PostWrapper>
-          <PostInput
-            type="text"
-            name="postalCode"
-            value={formData.postalCode}
+        {errors.username && <ErrorMessage>{errors.username}</ErrorMessage>}
+        <SubTWrapper>
+          <Subsubtitle>응원 팀 선택</Subsubtitle>
+          <StyledSelect
+            name="favoriteTeam"
+            value={formData.favoriteTeam}
             onChange={handleChange}
-            placeholder="우편번호"
-            readOnly
-          />
-          <PostButton type="button" onClick={openAddressModal}>
-            우편번호 검색
-          </PostButton>
-        </PostWrapper>
-        <Input
-          type="text"
-          name="address"
-          value={formData.address}
-          onChange={handleChange}
-          placeholder="주소"
-          readOnly
-        />
-        <Input
-          type="text"
-          name="detailedAddress"
-          value={formData.detailedAddress}
-          onChange={handleChange}
-          placeholder="상세주소"
-        />
-      </SubTWrapper>
-      <AllCheckWrapper>
-        <CheckWrapper>
-          <Checkoption>
-            <CheckCircle
-              checked={formData.agreements.all}
-              onClick={() => handleCheck("all")}
+            $isTeamPlaceholder={!formData.favoriteTeam}
+          >
+            <option value="" disabled>
+              응원하는 구단을 선택해 주세요
+            </option>
+            <option value="기아 타이거즈">기아 타이거즈</option>
+            <option value="삼성 라이온즈">삼성 라이온즈</option>
+            <option value="LG 트윈스">LG 트윈스</option>
+            <option value="두산 베어스">두산 베어스</option>
+            <option value="KT 위즈">KT 위즈</option>
+            <option value="SSG 랜더스">SSG 랜더스</option>
+            <option value="롯데 자이언츠">롯데 자이언츠</option>
+            <option value="한화 이글스">한화 이글스</option>
+            <option value="NC 다이노스">NC 다이노스</option>
+            <option value="키움 히어로즈">키움 히어로즈</option>
+          </StyledSelect>
+        </SubTWrapper>
+        {errors.favoriteTeam && (
+          <ErrorMessage>{errors.favoriteTeam}</ErrorMessage>
+        )}
+        <SubTWrapper>
+          <Subsubtitle>생년월일</Subsubtitle>
+          <StyledSelect2Wrapper>
+            <StyledSelect2
+              name="birthdate.year"
+              value={formData.birthdate.year}
+              onChange={handleChange}
+              $isBirthPlaceholder={formData.birthdate.year === ""}
             >
-              <img src={logon_check} alt="logon_check" />
-            </CheckCircle>
-            <CheckText
-              checked={formData.agreements.all}
-              onClick={() => handleCheck("all")}
+              <option value="" disabled>
+                년
+              </option>
+              {years.map((year) => (
+                <option key={year} value={year}>
+                  {year}
+                </option>
+              ))}
+            </StyledSelect2>
+            <StyledSelect2
+              name="birthdate.month"
+              value={formData.birthdate.month}
+              onChange={handleChange}
+              $isBirthPlaceholder={formData.birthdate.month === ""}
             >
-              필수 및 선택 사항에 모두 동의합니다.
-            </CheckText>
-          </Checkoption>
-        </CheckWrapper>
-        <Line />
-        <CheckWrapper>
-          <Checkoption>
-            <CheckCircle
-              checked={formData.agreements.required}
-              onClick={() => handleCheck("required")}
+              <option value="" disabled>
+                월
+              </option>
+              {months.map((month) => (
+                <option key={month} value={month}>
+                  {month}
+                </option>
+              ))}
+            </StyledSelect2>
+            <StyledSelect2
+              name="birthdate.date"
+              value={formData.birthdate.date}
+              onChange={handleChange}
+              $isBirthPlaceholder={formData.birthdate.date === ""}
             >
-              <img src={logon_check} alt="logon_check" />
-            </CheckCircle>
-            <CheckText
-              checked={formData.agreements.required}
-              onClick={() => handleCheck("required")}
-            >
-              [필수] 이용약관에 동의합니다. <span>*</span>
-            </CheckText>
-          </Checkoption>
-          <img
-            src={RBarrow}
-            alt="RBarrow"
-            onClick={() => openModal("required")}
-          />
-        </CheckWrapper>
-        <CheckWrapper>
-          <Checkoption>
-            <CheckCircle
-              checked={formData.agreements.privacy}
-              onClick={() => handleCheck("privacy")}
-            >
-              <img src={logon_check} alt="logon_check" />
-            </CheckCircle>
-            <CheckText
-              checked={formData.agreements.privacy}
-              onClick={() => handleCheck("privacy")}
-            >
-              [필수] 개인정보 수집 및 이용에 동의 합니다. <span>*</span>
-            </CheckText>
-          </Checkoption>
-          <img
-            src={RBarrow}
-            alt="RBarrow"
-            onClick={() => openModal("privacy")}
-          />
-        </CheckWrapper>
-        <CheckWrapper>
-          <Checkoption>
-            <CheckCircle
-              checked={formData.agreements.promotion}
-              onClick={() => handleCheck("promotion")}
-            >
-              <img src={logon_check} alt="logon_check" />
-            </CheckCircle>
-            <CheckText
-              checked={formData.agreements.promotion}
-              onClick={() => handleCheck("promotion")}
-            >
-              Rookie가 제공하는 이벤트 등 프로모션 안내 메일을 수신에
-              동의합니다.
-            </CheckText>
-          </Checkoption>
-        </CheckWrapper>
-        {errors.agreements && <ErrorMessage>{errors.agreements}</ErrorMessage>}
-      </AllCheckWrapper>
-      {errors.firebase && <ErrorMessage>{errors.firebase}</ErrorMessage>}
-
-      <LoginBtn type="submit" disabled={isLoading} $valid={isFormValid}>
-        {isLoading ? "로딩중..." : "회원가입"}
+              <option value="" disabled>
+                일
+              </option>
+              {dates.map((date) => (
+                <option key={date} value={date}>
+                  {date}
+                </option>
+              ))}
+            </StyledSelect2>
+          </StyledSelect2Wrapper>
+        </SubTWrapper>
+        {errors.birthdate && <ErrorMessage>{errors.birthdate}</ErrorMessage>}
+        <SubTWrapper>
+          <Subsubtitle>휴대폰 번호</Subsubtitle>
+          <StyledSelect2Wrapper>
+            <Input2
+              name="phoneNumber.part1"
+              value={formData.phoneNumber.part1}
+              onChange={handleChange}
+              placeholder="010"
+            />
+            <Input2
+              name="phoneNumber.part2"
+              value={formData.phoneNumber.part2}
+              onChange={handleChange}
+            />
+            <Input2
+              name="phoneNumber.part3"
+              value={formData.phoneNumber.part3}
+              onChange={handleChange}
+            />
+          </StyledSelect2Wrapper>
+          {errors.phoneNumber && (
+            <ErrorMessage>{errors.phoneNumber}</ErrorMessage>
+          )}
+        </SubTWrapper>
+      </AllInputWrapper>
+      <LoginBtn type="submit" $valid={isFormValid}>
+        다음
       </LoginBtn>
-
-      <LogonModal
-        isOpen={modalState.required}
-        closeModal={() => closeModal("required")}
-        contentType="required"
-      />
-      <LogonModal
-        isOpen={modalState.privacy}
-        closeModal={() => closeModal("privacy")}
-        contentType="privacy"
-      />
-
-      {isAddressModalOpen && (
-        <ModalOverlay onClick={closeAddressModal}>
-          <ModalContent onClick={(e) => e.stopPropagation()}>
-            <ModalCloseButton onClick={closeAddressModal}>×</ModalCloseButton>
-            <DaumPostcode onComplete={handleAddressComplete} />
-          </ModalContent>
-        </ModalOverlay>
-      )}
-    </Form>
+    </form>
   );
 };
