+++ src/pages/Login.jsx
@@ -227,12 +227,17 @@ const SnsLogoWrapper = styled.div`
 const LoginBtn = styled.button`
   width: 100%;
   height: 70px;
-  background: var(--grayE);
   border: none;
   border-radius: 4px;
   font-size: 2.4rem;
-  color: var(--grayC);
+  color: var(--light);
   cursor: pointer;
+  background: ${({ $valid }) => ($valid ? "var(--dark)" : "var(--grayC)")};
+  transition: background-color 0.3s;
+  &:hover {
+    background-color: ${({ $valid }) =>
+      $valid ? "var(--gray2)" : "var(--grayC)"};
+  }
   @media screen and (max-width: 1024px) {
     height: 56px;
     font-size: 1.6rem;
@@ -348,6 +353,7 @@ const Login = () => {
 
   const handlelogonclick = () => {
     navigate("/logon");
+    window.location.reload();
   };
 
   const isFormValid =
@@ -400,15 +406,7 @@ const Login = () => {
             <img src={kakao_talk} alt="kakao_talk" />
             <img src={naver_icon} alt="naver_icon" />
           </SnsLogoWrapper>
-          <LoginBtn
-            type="submit"
-            style={{
-              background:
-                isFormValid && !isLoading ? "var(--dark)" : "var(--grayE)",
-              color:
-                isFormValid && !isLoading ? "var(--light)" : "var(--grayC)",
-            }}
-          >
+          <LoginBtn type="submit" $valid={isFormValid}>
             {isLoading ? "로딩중..." : "로그인"}
           </LoginBtn>
         </Form>
